# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AppListItem:
    properties:
      app_id:
        description: "\b租户id"
        type: string
      id:
        type: integer
      name:
        description: 租户名称
        type: string
      qpd:
        description: WhiteIPs string `json:"white_ips"` // ip白名单，支持前缀匹配
        type: integer
      qps:
        type: integer
      secret:
        description: 密钥
        type: string
    type: object
...
  dto.GetAppListOutput:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.AppListItem'
        type: array
      total:
        type: integer
    type: object
...
  dto.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
...
host: localhost:8880
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /app/list:
    get:
      consumes:
      - application/json
      description: 获取租户列表
      operationId: /app/list
      parameters:
      - in: query
        name: info
        type: string
      - example: 2
        in: query
        name: page_no
        required: true
        type: integer
      - example: 10
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetAppListOutput'
              type: object
      summary: 获取租户列表
      tags:
      - 租户接口
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"
...